refactor(structure): reorganize project directories
- Move abstraction classes to core/interfaces/
- Elevate config/ to root level
- Group helpers in utils/helpers/
- Remove empty types/ directory
- Update import paths
