+++ src/adapters/akamai/akamaiAdapter.js
@@ -1,8 +1,8 @@
 // akamaiAdapter.js
 
-import * as optlyHelper from '../../utils/optimizelyHelper';
-import * as cookieDefaultOptions from '../../utils/config/cookieOptions';
-import defaultSettings from '../../utils/config/defaultSettings';
+import * as optlyHelper from '../../utils/helpers/optimizelyHelper';
+import * as cookieDefaultOptions from '../../config/cookieOptions';
+import defaultSettings from '../../config/defaultSettings';
 import EventListeners from '../../core/providers/events/eventListeners';
 
 /**
