+++ src/adapters/cloudflare/cloudflareKVInterface.js
@@ -10,8 +10,8 @@
  * - delete(key) - Deletes a key from the KV store.
  */
 
-import { logger } from '../../utils/optimizelyHelper';
-import { KVStoreAbstractInterface } from '../../utils/abstraction-classes/kvStoreAbstractInterface';
+import { logger } from '../../utils/helpers/optimizelyHelper';
+import { KVStoreAbstractInterface } from '../../core/interfaces/kvStoreAbstractInterface';
 
 /**
  * Class representing the Cloudflare KV store interface.
