+++ src/adapters/fastly/fastlyKVInterface.js
@@ -9,8 +9,8 @@
  * - delete(key) - Deletes a key from the Fastly EdgeWorkers KV store.
  */
 
-import { logger } from '../../utils/optimizelyHelper.js';
-import { KVStoreAbstractInterface } from '../../utils/abstraction-classes/kvStoreAbstractInterface';
+import { logger } from '../../utils/helpers/optimizelyHelper.js';
+import { KVStoreAbstractInterface } from '../../core/interfaces/kvStoreAbstractInterface';
 
 /**
  * Class representing the Fastly EdgeWorkers KV store interface.
