+++ src/core/api/handlers/variationChanges.js
@@ -2,8 +2,8 @@
  * @module VariationChanges
  */
 
-import { logger } from '../../../utils/optimizelyHelper.js';
-import defaultSettings from '../../../utils/config/defaultSettings.js';
+import { logger } from '../../../utils/helpers/optimizelyHelper.js';
+import defaultSettings from '../../../config/defaultSettings.js';
 
 /**
  * Fetches and updates the variation changes from the Optimizely API.
