+++ src/core/interfaces/abstractContext.js
