+++ src/core/interfaces/abstractRequest.js
