+++ src/core/interfaces/abstractResponse.js
