+++ src/core/interfaces/kvStoreAbstractInterface.js
