+++ src/core/providers/coreLogic.js
@@ -2,11 +2,11 @@
  * @module coreLogic
  */
 
-import * as optlyHelper from '../../utils/optimizelyHelper';
-import RequestConfig from '../../utils/config/requestConfig';
-import defaultSettings from '../../utils/config/defaultSettings';
-import { logger } from '../../utils/optimizelyHelper';
-import { AbstractionHelper } from '../../utils/abstractionHelper';
+import * as optlyHelper from '../../utils/helpers/optimizelyHelper';
+import RequestConfig from '../../config/requestConfig';
+import defaultSettings from '../../config/defaultSettings';
+import { logger } from '../../utils/helpers/optimizelyHelper';
+import { AbstractionHelper } from '../../utils/helpers/abstractionHelper';
 import EventListeners from './events/eventListeners';
 
 /**
