+++ src/core/providers/optimizelyProvider.js
@@ -2,10 +2,10 @@
  * @module OptimizelyProvider
  */
 
-import * as optlyHelper from '../../utils/optimizelyHelper';
-import { logger } from '../../utils/optimizelyHelper';
-import { AbstractionHelper } from '../../utils/abstractionHelper';
-import defaultSettings from '../../utils/config/defaultSettings';
+import * as optlyHelper from '../../utils/helpers/optimizelyHelper';
+import { logger } from '../../utils/helpers/optimizelyHelper';
+import { AbstractionHelper } from '../../utils/helpers/abstractionHelper';
+import defaultSettings from '../../config/defaultSettings';
 import UserProfileService from './userProfileService';
 import EventListeners from './events/eventListeners';
